Moved all the OnNotify values into the one global enum.

+++ src/common/Text/GHtml.cpp
@@ -4076,7 +4076,7 @@ void GHtmlTableLayout::DeallocatePx(int StartCol, int Cols, int MaxPx)
 	int TrimPx = TotalPx - MaxPx;
 	GArray<ColInfo> Inf;
 	int HalfMax = MaxPx >> 1;
-	int Interesting = 0;
+	unsigned Interesting = 0;
 	int InterestingPx = 0;
 	
 	for (int x=StartCol; x<StartCol+Cols; x++)
@@ -4099,7 +4099,7 @@ void GHtmlTableLayout::DeallocatePx(int StartCol, int Cols, int MaxPx)
 	{
 		ColInfo &ci = Inf[i];
 		double r = (double)ci.Px / InterestingPx;
-		int DropPx = r * TrimPx;
+		int DropPx = (int) (r * TrimPx);
 		if (DropPx < MinCol[ci.Idx])
 			MinCol[ci.Idx] -= DropPx;
 		else
